+++ gems/quilt_rails/lib/quilt_rails.rb
@@ -8,3 +8,4 @@ require "quilt_rails/logger"
 require "quilt_rails/configuration"
 require "quilt_rails/react_renderable"
 require "quilt_rails/trusted_ui_server_csrf_strategy"
+require "quilt_rails/monkey_patches/active_support_reloader" if Rails.env.development?